backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy
psycopg[binary]
alembic
python-jose[cryptography]
passlib[argon2]
python-multipart
email-validator
pydantic-settings
apscheduler
resend
google-auth
reportlab
